MXS-839: Detect multi-master topologies with mysqlmon
The mysqlmon now supports proper detection of multi-master topologies by
building a directed graph out of the monitored server. If cycles are found
from this graph, they are assigned a master group ID. All servers with a
master group ID will receive the Master status unless they have
`@@read_only` enabled.

This new functionality can be enabled with the 'multimaster' boolean
parameter. Currently stale master server detection is not supported but
this will be addressed in a later commit.